Fix compilation errors in new game files
- mpu6050.cpp: Replace brace initializer for IMUCalibration with
  zero-initialization using static storage duration defaults
- game_rhythm.h: Remove duplicate RhythmGameState struct definition;
  struct is already defined in types.h
- game_rhythm.cpp: Add forward declaration for rhythmGameStartRound()
  so it can be called before its definition; simplify global init
- game_balance.cpp: Cast BALANCE_HARD enum to int for min() call
  to resolve type deduction failure with mixed enum/int args
- ui_play_rhythm.cpp: Replace undefined FreeSans font references
  with setTextSize() matching the codebase's built-in GFX font
- ui_play_balance.cpp: Same font fix; also add proper erase-and-
  redraw logic to avoid full screen repaints on ball movement
- types.h: Use = {} instead of = "" for char array default init

+++ game_balance.cpp
@@ -246,7 +246,7 @@ void balanceGameCheckWinCondition() {
       balanceGame->level = 1;  // Loop back
     }
 
-    int nextDifficulty = min(BALANCE_HARD, balanceGame->difficulty + 1);
+    int nextDifficulty = min((int)BALANCE_HARD, balanceGame->difficulty + 1);
     if (balanceGame->level == 1 && balanceGame->difficulty == BALANCE_HARD) {
       nextDifficulty = BALANCE_EASY;  // Reset difficulty on cycle
     }
